Idea 1 poc/Vertex AI Setup/requirements.txt
google-genai
google-cloud-bigquery
sentence-transformers
chromadb
numpy
fastapi
uvicorn[standard]
